// dojo/src/components/constants/types.ts
export interface BodyCheck {
  id: number;
  temp_id: string;
  check_date: string;
  overall_status: 'pass' | 'fail';
  color_vision: string;
  eye_movement: string;
  fingers_functionality: string;
  hand_deformity: string;
  joint_mobility: string;
  hearing: string;
  bending_ability: string;
  additional_checks: {
    description: string;
    status: 'pass' | 'fail' | '';
  }[];
  notes: string;
}
// ...
export interface OrientationFeedback {
  id: number;
  pay_code: string;
  card_no: string;
  sex: string;
  birth_date: string;
  guardian_name: string;
  department: string;
  section: string;
  desig_category: string;
  joining_date: string;
  user: number;
}

// Full User with Relations


export interface User {
  id: number;
  first_name: string;
  last_name: string;
  temp_id: string;
  email: string;
  phone_number: string;
  sex: string;
  created_at: string;
  updated_at: string;
  is_active: boolean;
  body_checks: BodyCheck[];
  orientation_feedbacks: OrientationFeedback[];
  photo?: string; 
  is_added_to_master: boolean;
  added_to_master_at: string | null;
  aadharNumber: string;
  employment_type: string;
  hasExperience: boolean;
  experienceYears: number;
  companyOfExperience: string;
}
// ...
export interface OrientationFeedbackModalProps {
  user: User;
  onClose: () => void;
  onSave: () => void;
}